Add files via upload
Finally managed to get a loop working for adding the floor which I can alter to be any length I want. Works perfectly with my modified ground texture + allows for big highscores ;) 

Alter the length of the floor by changing the FOR loop in "settings.py" to be X length (default is 100) .

To add any actual platforms that aren't making up the ground you must append after the main loop in "settings.py".